moved 1st implentation of Resolution to playground
