+++ cmsl1t/playground/resolution.py
